+++ packages/@headlessui-react/src/hooks/use-is-mounted.ts
@@ -1,9 +1,11 @@
 import { useRef, useEffect } from 'react'
 
 export function useIsMounted() {
-  let mounted = useRef(true)
+  let mounted = useRef(false)
 
   useEffect(() => {
+    mounted.current = true
+
     return () => {
       mounted.current = false
     }