+++ packages/@headlessui-react/src/hooks/use-window-event.ts
@@ -1,12 +1,19 @@
-import { useEffect } from 'react'
+import { useEffect, useRef } from 'react'
 
 export function useWindowEvent<TType extends keyof WindowEventMap>(
   type: TType,
   listener: (this: Window, ev: WindowEventMap[TType]) => any,
   options?: boolean | AddEventListenerOptions
 ) {
+  let listenerRef = useRef(listener)
+  listenerRef.current = listener
+
   useEffect(() => {
-    window.addEventListener(type, listener, options)
-    return () => window.removeEventListener(type, listener, options)
-  }, [type, listener, options])
+    function handler(event: WindowEventMap[TType]) {
+      listenerRef.current.call(window, event)
+    }
+
+    window.addEventListener(type, handler, options)
+    return () => window.removeEventListener(type, handler, options)
+  }, [type, options])
 }